+++ lib/surface/ast.ex
@@ -66,8 +66,8 @@ defmodule Surface.AST.Meta do
       * `:line_offset` - the line offset from the caller's line to the start of this source
       * `:caller` - a Macro.Env struct representing the caller
   """
-  @derive {Inspect, only: [:line, :module, :node_alias, :file]}
-  defstruct [:line, :module, :node_alias, :line_offset, :file, :caller]
+  @derive {Inspect, only: [:line, :module, :node_alias, :file, :checks]}
+  defstruct [:line, :module, :node_alias, :line_offset, :file, :caller, :checks]
 
   @type t :: %__MODULE__{
           line: non_neg_integer(),
@@ -75,7 +75,8 @@ defmodule Surface.AST.Meta do
           module: atom(),
           node_alias: binary() | nil,
           caller: Macro.Env.t(),
-          file: binary()
+          file: binary(),
+          checks: Keyword.t(boolean())
         }
 
   def quoted_caller_cid(meta) do