+++ test/context_change_tracking_test.exs
@@ -41,6 +41,7 @@ defmodule Surface.ContextChangeTrackingTest do
     alias Surface.CheckUpdated
 
     data count, :integer, default: 0
+    data test_pid, :integer
 
     def mount(_params, %{"test_pid" => test_pid}, socket) do
       {:ok, assign(socket, test_pid: test_pid)}