+++ test/slot_change_tracking_test.exs
@@ -23,6 +23,7 @@ defmodule Surface.SlotChangeTrackingTest do
     alias Surface.CheckUpdated
 
     data count, :integer, default: 0
+    data test_pid, :integer
 
     def mount(_params, %{"test_pid" => test_pid}, socket) do
       {:ok, assign(socket, test_pid: test_pid)}